+++ app/filePrepare/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+filePrepare = Blueprint('filePrepare', __name__)
+
+from . import views
