+++ app/main/forms.py
@@ -1,10 +1,8 @@
 from flask_wtf import FlaskForm
-from datetime import date
-from wtforms import SubmitField, DateField
+from wtforms import StringField, SubmitField
 from wtforms.validators import DataRequired
 
 
-class DownloadForm(FlaskForm):
-    dateFrom = DateField('Pēdējās datu atjaunošanas datums',
-                         default=date.today, format='%Y-%m-%d', validators=[DataRequired()])
-    submit = SubmitField('Iesniegt')
+class NameForm(FlaskForm):
+    name = StringField('What is your name?', validators=[DataRequired()])
+    submit = SubmitField('Submit')
