+++ app/templates/base.html
@@ -23,7 +23,9 @@
         <div class="navbar-collapse collapse">
             <ul class="nav navbar-nav">
                 <li><a href="{{ url_for('main.index') }}">Sākums</a></li>
-                <li><a href="{{ url_for('main.download') }}">Medikamentu datu atjaunošana</a></li>
+                {% if current_user.is_authenticated %}
+                <li><a href="{{ url_for('filePrepare.download') }}">Medikamentu datu atjaunošana</a></li>
+                {% endif %}
             </ul>
             <ul class="nav navbar-nav navbar-right">
                 {% if current_user.is_authenticated %}
