+++ app/templates/filePrepare/checkMedication.html
@@ -1,0 +1,9 @@
+{% extends "base.html" %}
+
+{% block title %}ZRApp - pārbaudīt medikamentus{% endblock %}
+
+{% block page_content %}
+<div class="page-header">
+    <h1>Pārbaudīt medikamentus</h1>
+</div>
+{% endblock %}
